Added traf workflow

+++ workflows/multall-training/multall_flow.py
@@ -1,5 +1,5 @@
 """
-@brief This runs a multallt set to produce training/validation data for training a QML/ML
+@brief This runs a multall set to produce training/validation data for training a QML/ML
 
 """
 
@@ -15,7 +15,7 @@ import datetime
 from typing import List, Set, Callable, Tuple, Dict, Any
 from qbitbridge.vqpubase import HybridQuantumWorkflowBase
 import asyncio
-from prefect import task, flow
+from prefect import task, flow, concurrency
 from prefect.logging import get_run_logger
 import numpy as np
 import subprocess
@@ -113,13 +113,15 @@ def process_and_write_multall_output(
         data : Dict,
         output_name : str) -> None:
     """process and save relevant output from all multall runs"""
-    for key in results.keys():
+    for key in rawresults.keys():
         i = data["pname"].index(key)
-        data["TOTAL TO TOTAL ISENTROPIC EFFICIENCY"][i]=float(results[key]["TOTAL TO TOTAL ISENTROPIC EFFICIENCY"][0])
+        data["TOTAL TO TOTAL ISENTROPIC EFFICIENCY"][i]=float(rawresults[key]["TOTAL TO TOTAL ISENTROPIC EFFICIENCY"][0])
+        data["EXIT STAGNATION TEMPERATURE"][i]=float(rawresults[key]["INLET AND EXIT STAGNATION TEMPERATURES"][1])
+        data["EXIT STAGNATION PRESSURE"][i]=float(rawresults[key]["INLET AND EXIT STAGNATION PRESSURES"][1])
     df1 = pandas.DataFrame(data)
     # need to figure out best way of saving results
     # for now we produce a csv file with results 
-    df1.to_csv(f"{output_name}.csv")    
+    df1.to_csv(f"{output_name}.csv")
 
 
 @task
@@ -135,11 +137,12 @@ def run_multall(
     """
     Task running the process to get multall output given a meangen input
     """
+    # with concurrency("data-processing", occupy=256):
     if outdir[-1] != "/":
         outdir+="/"
     logger = get_run_logger()
     logger.info(f"Running multall for {param_set_name}")
-    os.mkdir(outdir)
+    os.makedirs(outdir, exist_ok=True)
 
     info: Dict[str, str] = {}
     if meangen_template is not None:
@@ -157,25 +160,30 @@ def run_multall(
         multallexecs = ["stagen", "multall"]
 
     for cmd in multallexecs:
-        scriptname = f"{outdir}run_{cmd}.sh"
-        with open(scriptname, 'w') as f:
-            f.write(f"#!/bin/bash\n")
-            f.write(f"cd {outdir} \n")
-            f.write(f"export MEANGEN_ARGS=\"meangen_{param_set_name}.in meangen_{param_set_name} \"\n")
-            f.write(f"export STAGEN_ARGS=\"meangen_{param_set_name}_stagen.dat stagen_out_{param_set_name} \"\n")
-            f.write(f"export MULTALL_ARGS=\"stagen_out_{param_set_name}_new.dat multall_{param_set_name} \"\n")
-            f.write(f"{execdir}/{cmd}\n")
-        os.chmod(scriptname, 0o755)
-        logger.info(f"Running {cmd} ...")
-        process = subprocess.run(
-            [scriptname],
-            capture_output=True,
-            text=True,
-        )
-        # do some post processing of output if required
-        info[cmd] = process.stdout
-        with open(f"{outdir}{cmd}.log", "w") as f:
-            f.write(info[cmd])
+        if os.path.isfile(f"{outdir}{cmd}.log"):
+            logger.info(f"Skipping {cmd} as log file exists")
+            with open(f"{outdir}{cmd}.log", "r") as f:
+                info[cmd] = "".join(f.readlines())
+        else:
+            scriptname = f"{outdir}run_{cmd}.sh"
+            with open(scriptname, 'w') as f:
+                f.write(f"#!/bin/bash\n")
+                f.write(f"cd {outdir} \n")
+                f.write(f"export MEANGEN_ARGS=\"meangen_{param_set_name}.in meangen_{param_set_name} \"\n")
+                f.write(f"export STAGEN_ARGS=\"meangen_{param_set_name}_stagen.dat stagen_out_{param_set_name} \"\n")
+                f.write(f"export MULTALL_ARGS=\"stagen_out_{param_set_name}_new.dat multall_{param_set_name} \"\n")
+                f.write(f"{execdir}/{cmd}\n")
+            os.chmod(scriptname, 0o755)
+            logger.info(f"Running {cmd} ...")
+            process = subprocess.run(
+                [scriptname],
+                capture_output=True,
+                text=True,
+            )        
+            # do some post processing of output if required
+            info[cmd] = process.stdout
+            with open(f"{outdir}{cmd}.log", "w") as f:
+                f.write(info[cmd])
         logger.info(f"Finished running {cmd}")
 
     # parse the info from multall
@@ -218,11 +226,12 @@ async def multall_workflow(
     meangen_template: str | None,
     stagen_template: str | None, 
     params: Dict[str, Any],
+    baseoutput_dir: str = "./",
     output_name : str = "multall_runs", 
     run_name : str = "run",
     multallexecs: List[str] = ["meangen", "stagen", "multall"],
     execdir : str = "/software/projects/pawsey0001/pelahi/pawsey-uptake-project-quantum-umelb/CFD/multall/bin/",
-    hash_length : int = 6,
+    hash_length : int = 10,
     resultkeys : List[str] = [
         "INLET STAGNATION PRESSURE",
         "EXIT STAGNATION PRESSURE",
@@ -239,6 +248,7 @@ async def multall_workflow(
         "MID STATIC TEMPERATURES",
         "EXIT STATIC TEMPERATURES",
     ],
+    max_task_submissions : int = 256,
     date: datetime.datetime = datetime.datetime.now(),
 ) -> None:
     """Flow for running multall 
@@ -261,10 +271,12 @@ async def multall_workflow(
         data[key] = [None for i in range(len(param_values))]
     for key in resultkeys:
         data[key] = [None for i in range(len(param_values))]
+    logger.info(f"{data.keys()}")
 
     # submit the task and wait for results
     counter = 0
     futures = {}
+    results = {}
     for param_vals in param_values:
         pset = list()
         pname = list()
@@ -278,8 +290,8 @@ async def multall_workflow(
         sha256_hash.update(pname.encode('utf-8'))
         # take the first n from the hash
         hex_digest = sha256_hash.hexdigest()[:hash_length]
-        outdir = f"{run_name}-{hex_digest}/"
-        if not os.path.isdir(outdir):
+        outdir = f"{baseoutput_dir}/{run_name}-{hex_digest}/"
+        if not os.path.isfile(f"{outdir}/multall.log"):
             futures[pname] = run_multall.submit(
                     param_set_name=pname,
                     param_set=pset,
@@ -290,82 +302,112 @@ async def multall_workflow(
                     outdir=outdir,
 
                 )
-            data["pname"][counter]=pname
-            for p in param_names:
-                data[p][counter]=param_vals[i]
-            data["outdir"][counter]=outdir
+        else:
+            futures[pname] = None
+            results[pname] = run_multall.fn(
+                    param_set_name=pname,
+                    param_set=pset,
+                    multallexecs=multallexecs,
+                    execdir=execdir,
+                    meangen_template=meangen_template,
+                    stagen_template=stagen_template,
+                    outdir=outdir,
+
+            )
+        data["pname"][counter]=pname
+        for p in param_names:
+            data[p][counter]=param_vals[i]
+        data["outdir"][counter]=outdir
         counter += 1
-    results = {}
+        if counter % max_task_submissions == 0:
+            logger.info(f"Pausing task submission for 10 seconds to not overload database")
+            sleep(10)
     # get the results from running the multall tasks
     for key in futures.keys():
-        results[key] = futures[key].result()
-    process_and_write_multall_output(results, data, output_name)
+        if futures[key] is not None:
+            results[key] = futures[key].result()
+    logger.info(f"{results}")
+    # logger.info(results[key])
+    process_and_write_multall_output(results, data, f"{baseoutput_dir}/{output_name}")
     logger.info("Finished multall CPU flow")
 
 
 def wrapper_to_async_flow(
-    meangen_template: str = "./meangen.in.template",
-    stagen_template: str | None = None, #"./stagen.in.template",
-    params: Dict[str, Any] = {
-        "FLOW_ANGLE": np.arange(0, 10, 5),
-        "FLOW_SPEED": None,
-        "BLADE_SHAPE": None,
-    },
-    yaml_template: str | None = None,
-    script_template: str | None = None,
-    cluster: str | None = None,
+    args: Any,
 ) -> None:
     """
     Run the multall runner
     """
-    if yaml_template == None:
-        yaml_template = f"{os.path.dirname(os.path.abspath(__file__))}/../../qbitbridge/workflow/qb-vqpu/remote_vqpu_template.example.yaml"
-    if script_template == None:
-        script_template = f"{os.path.dirname(os.path.abspath(__file__))}/../../qbitbridge/workflow/qb-vqpu/vqpu_template.example.sh"
-    if cluster == None:
-        cluster = "multall-setonix-pypath"
+    
+    if args.params_file is not None:
+        import json
+
+        if not os.path.exists(args.params_file):
+            raise ValueError(f"Parameter file {args.params_file} does not exist")
+        with open(args.params_file, "r") as f:
+            params = json.load(f)
+    else:
+        params: Dict[str, Any] = {
+            "FLOW_ANGLE": np.arcsin(np.linspace(np.sin(-70/180*np.pi), np.sin(70/180*np.pi), num=255, endpoint=True))/np.pi*180,
+            "FLOW_SPEED": None,
+            "BLADE_SHAPE": None,
+        }
+
     myflow = HybridQuantumWorkflowBase(
-        cluster=cluster,
+        cluster=args.cluster,
         vqpu_ids=[1, 2, 3, 16],
-        vqpu_template_yaml=yaml_template,
-        vqpu_template_script=script_template,
+        vqpu_template_yaml=args.yaml_template,
+        vqpu_template_script=args.script_template,
         eventloc=f"{os.path.dirname(os.path.abspath(__file__))}/events/",
     )
 
-    if stagen_template is not None:
-        if not os.path.exists(stagen_template):
-            raise ValueError(f"Stagen file {stagen_template} does not exist")
+    if args.stagen_template is not None:
+        if not os.path.exists(args.stagen_template):
+            raise ValueError(f"Stagen file {args.stagen_template} does not exist")
         # just bypass meangen if stagen template is used 
-        meangen_template = None
+        args.meangen_template = None
 
-    if meangen_template is not None:
-        if not os.path.exists(meangen_template):
-            raise ValueError(f"Meangen file {meangen_template} does not exist")
+    if args.meangen_template is not None:
+        if not os.path.exists(args.meangen_template):
+            raise ValueError(f"Meangen file {args.meangen_template} does not exist")
 
     asyncio.run(
-        multall_workflow.with_options(task_runner=myflow.gettaskrunner("cpu"))(
-            myqpuworkflow=myflow,
-            meangen_template=meangen_template,
-            stagen_template=stagen_template, 
+        multall_workflow.with_options(task_runner=myflow.gettaskrunner("cpu-single"))(
+            myqpuworkflow = myflow,
+            meangen_template = args.meangen_template,
+            stagen_template = args.stagen_template, 
             params = params,
+            baseoutput_dir = args.output_dir,
+            run_name = args.run_name,
+            output_name = args.output_name,
         )
     )
 
+def multall_argparse():
+    import argparse
+    # Create the argument parser
+    parser = argparse.ArgumentParser(description="Multall workflow arguments.")
+
+    # Define arguments
+    parser.add_argument("--output_dir", type = str, default = "./", help="output path for results")
+    parser.add_argument("--yaml_template", type=str, 
+                        default=f"{os.path.dirname(os.path.abspath(__file__))}/../../qbitbridge/workflow/qb-vqpu/remote_vqpu_template.example.yaml", 
+                        help="yaml file for the vqpu")
+    parser.add_argument("--script_template", type=str, 
+                        default=f"{os.path.dirname(os.path.abspath(__file__))}/../../qbitbridge/workflow/qb-vqpu/vqpu_template.example.sh",
+                        help="script template file for the vqpu")
+    parser.add_argument("--cluster", type=str, default="multall-setonix-pypath", help="cluster name for running workflow")
+    parser.add_argument("--meangen_template", type=str, default="./meangen.in.template", help="template for meangen input file")
+    parser.add_argument("--stagen_template", type=str, default=None, help="template for stagen input file")
+    parser.add_argument("--run_name", type=str, default="run", help="name for the run")
+    parser.add_argument("--output_name", type=str, default="multall_runs", help="name for the output file")
+    parser.add_argument("--params_file", type=str, default=None, help="json file containing the parameters to run")
+
+    # Parse the arguments
+    args = parser.parse_args()
+    return args
+
 if __name__ == "__main__":
-    yaml_template = None
-    script_template = None
-    cluster = None
-    res = [i for i in sys.argv if re.findall("--yaml=", i)]
-    if len(res) > 0:
-        yaml_template = res[0].split("=")[1]
-    res = [i for i in sys.argv if re.findall("--script=", i)]
-    if len(res) > 0:
-        script_template = res[0].split("=")[1]
-    res = [i for i in sys.argv if re.findall("--cluster=", i)]
-    if len(res) > 0:
-        cluster = res[0].split("=")[1]
-    wrapper_to_async_flow(
-        yaml_template=yaml_template,
-        script_template=script_template,
-        cluster=cluster,
-    )
+
+    args = multall_argparse()
+    wrapper_to_async_flow(args=args)
